requirements.txt
pandas
numpy
torch
transformers
sentence-transformers
Flask
faiss-cpu
selenium
torch